[api] merge organization to archive model and repo into organization for admin
