+++ api/lib/domain/usecases/archive-organization.js
@@ -1,14 +1,4 @@
-const OrganizationToArchive = require('../models/OrganizationToArchive');
-
-module.exports = async function archiveOrganization({
-  organizationId,
-  userId,
-  organizationToArchiveRepository,
-  organizationForAdminRepository,
-}) {
-  const organizationToArchive = new OrganizationToArchive({ id: organizationId });
-  organizationToArchive.archive({ archivedBy: userId });
-  await organizationToArchiveRepository.save(organizationToArchive);
-
+module.exports = async function archiveOrganization({ organizationId, userId, organizationForAdminRepository }) {
+  await organizationForAdminRepository.archive({ id: organizationId, archivedBy: userId });
   return await organizationForAdminRepository.get(organizationId);
 };
